--- test/test_gettranslatedtokens.py
@@ -1,77 +1,0 @@
-import pytest
-import requests
-import json
-
-@pytest.fixture
-def supply_url():
-	return "https://stagingapi.autographamt.com"
-
-
-@pytest.fixture
-def get_adm_accessToken():
-	email = "alex@yopmail.com"
-	password = "1189"
-	url = "https://stagingapi.autographamt.com/v1/auth"
-	data = {'email':email,
-			'password':password}
-	resp = requests.post(url, data=data)
-	respobj = json.loads(resp.text)
-	token = respobj['accessToken']
-
-	return token
-
-@pytest.fixture
-def get_supAdmin_accessToken():
-	email = 'savitha.mark@bridgeconn.com'
-	password = '221189'
-	url = "https://stagingapi.autographamt.com/v1/auth"
-	data = {'email':email,
-			'password':password}
-	resp = requests.post(url, data=data)
-	respobj = json.loads(resp.text)
-	token = respobj['accessToken']
-
-	return token
-
-@pytest.fixture
-def get_trans_accessToken():
-	email = 'ag2@yopmail.com'
-	password = '1189'
-	url = "https://stagingapi.autographamt.com/v1/auth"
-	data = {'email':email,
-			'password':password}
-	resp = requests.post(url, data=data)
-	respobj = json.loads(resp.text)
-	token = respobj['accessToken']
-	return token
-
-
-def test_gettranslatedtokensup(supply_url,get_supAdmin_accessToken):
-	url = supply_url + '/v1/info/translatedtokens'
-	resp = requests.get(url,headers={'Authorization': 'bearer {}'.format(get_supAdmin_accessToken)})
-	j = json.loads(resp.text)
-	assert resp.status_code == 200, resp.text
-	assert isinstance(j,list), j
-	assert 'projectId' in j[0], j[0]
-	assert 'projectName' in j[0], j[0]
-	
-
-def test_gettranslatedtokenad(supply_url,get_adm_accessToken):
-	url = supply_url + '/v1/info/translatedtokens'
-	resp = requests.get(url,headers={'Authorization': 'bearer {}'.format(get_adm_accessToken)})
-	j = json.dumps(resp.text)
-	assert resp.status_code == 200, resp.text
-	# assert isinstance(j,list), j
-	# assert 'projectId' in j[0], j[0]
-	# assert 'projectName' in j[0], j[0]
-   
-   
-def test_gettranslatedtokentr(supply_url,get_trans_accessToken):
-	url = supply_url + '/v1/info/translatedtokens'
-	resp = requests.get(url,headers={'Authorization': 'bearer {}'.format(get_trans_accessToken)})
-	j = json.dumps(resp.text)
-	assert resp.status_code == 200, resp.text
-	# assert isinstance(j,list), j
-	# assert 'projectId in j[0], j[0]
-	# assert 'projectName' in j[0], j[0]
-
